PathTree extract node types into separate classes
Cleaning up here before moving into bug fixing: the original pass here
worked, but IMHO wasn't a great separation of concerns.

This separates the concerns more: me have `DirNode` & `FileNode` types,
which represent actual nodes in the tree, and `PathTree`, which now
pretty much is just a reference to a root node, and provides a
convenient interface for operating on the tree represented by the root
node.

I think this makes the similarities & differences between how dirs/files
are treated clearer.
